update: withServiceWorker convert to tsx
